Sandbox: Separate desktop notifications from tools/terkin.py
This fixes the nasty "ModuleNotFoundError: No module named ‘netaddr’" 
dependency bug on regular sandbox operations when not running
the Terkin Agent at all.

On top of that, add gracefulness to the invocation of "make notify"
to continue the task even if desktop notifications fail, whatever
reason that might have.
