+++ tools/terkin.mk
@@ -50,7 +50,7 @@ restart-device-http:
 	$(eval ip_address := $(mcu_port))
 
 	@# Notify user about the power cycling.
-	@$(MAKE) notify status=INFO status_ansi="$(INFO)" message="Restarting device at IP address $(ip_address) using HTTP API"
+	@$(MAKE) notify status=INFO status_ansi="$(INFO)" message="Restarting device at IP address $(ip_address) using HTTP API."
 
 	@# Send restart command to HTTP API
 	@# TODO: If this fails, maybe reset automatically using the serial interface.
@@ -73,3 +73,38 @@ restart-device-http:
 
 # Whether to use mpy-cross
 $(eval mpy_cross     := ${MPY_CROSS})
+
+
+# ============
+# Terkin Agent
+# ============
+
+# About: The Terkin Agent assists with device discovery and maintenance.
+
+
+## Setup the Terkin Agent.
+setup-terkin-agent:
+	$(pip3) install --pre --requirement requirements-terkin-agent.txt
+
+## Run the Terkin Agent, e.g. "make terkin-agent action=maintain".
+terkin-agent: setup-terkin-agent
+	sudo $(python3) -m tools.terkin $(action) $(macs)
+
+
+## Load the MiniNet module to the device and start a WiFi access point.
+provide-wifi: check-mcu-port
+	@$(rshell) $(rshell_options) --quiet cp lib/mininet.py /flash/lib
+	@$(rshell) $(rshell_options) --quiet repl "~ from mininet import MiniNet ~ MiniNet().activate_wifi_ap()"
+	@echo
+
+## Load the MiniNet module to the device and start a WiFi STA connection.
+connect-wifi: check-mcu-port
+	@$(rshell) $(rshell_options) --quiet cp lib/mininet.py /flash/lib/mininet_wip.py
+	@$(rshell) $(rshell_options) --quiet repl "~ from mininet_wip import MiniNet ~ MiniNet().connect_wifi_sta('$(ssid)', '$(password)')"
+	@echo
+
+## Load the MiniNet module to the device and get IP address.
+ip-address: check-mcu-port
+	@$(rshell) $(rshell_options) --quiet cp lib/mininet.py /flash/lib
+	@$(rshell) $(rshell_options) --quiet repl "~ from mininet import MiniNet ~ print(MiniNet().get_ip_address()) ~"
+	@echo
